DOM helpers moved to their own file
