+++ src/core/core.interaction.js
@@ -17,7 +17,7 @@ function getRelativePosition(e, chart) {
 		};
 	}
 
-	return helpers.getRelativePosition(e, chart);
+	return helpers.dom.getRelativePosition(e, chart);
 }
 
 /**
